+++ database/migrations/2020_11_13_214818_create_users_events_table.php
@@ -15,8 +15,8 @@ class CreateUsersEventsTable extends Migration
     {
         Schema::create('users_events', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('event_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('event_id');
             $table->timestamps();
         });
     }
